add confirm delete page

+++ products/views.py
@@ -125,10 +125,15 @@ def edit_product(request, product_id):
 def delete_product(request, product_id):
     """ Delete a product from the store """
     if not request.user.is_superuser:
+        # Redirect back to homepage if not a superuser
         messages.error(request, 'Sorry, only store owners can do that.')
         return redirect(reverse('home'))
-    
+
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+
+    if request.method == 'POST':
+        product.delete()
+        messages.success(request, f'{product.name} deleted!')
+        return redirect(reverse('products'))
+
+    return render(request, 'products/confirm_delete.html')
